FontRendererAPI: Move up into api package
