+++ src/main/java/li/cil/tis3d/api/FontRendererAPI.java
@@ -1,6 +1,5 @@
-package li.cil.tis3d.api.detail;
+package li.cil.tis3d.api;
 
-import li.cil.tis3d.api.ClientAPI;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.render.VertexConsumer;
